Update README.md
What is MEAN?
Each letter in the word MEAN has some specific meaning. Here, 'M' stands for MongoDB, 'E' stands for Express, 'A' stands for Angular, and 'N' stands for Node.js. It is one of the most popular stacks used for developing the full stack application. Let's understand the basic idea behind the mean stack.

As we can observe in the above figure that there is a front-end app, back-end app and database. The front-end app can be developed using either Angular.js or React.js, and back-end can be developed using Node.js, which is further connected to the MongoDB database. The front-end app and back-end app communicate with each other through the RestAPI. The back-end app exposes the RestAPI endpoints, whereas the front-end app consumes the RestAPI endpoints.
Let's look at each component of Mean.js one by one.

Node.js
Node.js
is an open-source platform and provides a runtime environment for executing the javascript code. It is mainly used for building the back-end application. Since there are two types of apps, such as web apps and mobile apps, where web apps run on the browser and mobile apps run on mobile devices. Both web app and mobile app are the client apps to which the user interacts. These apps require to communicate with the backend services to store the data, send emails, push notifications. Node.js is an ideal platform to build highly scalable, data-intensive, and real-time applications. It can be used for agile development and highly-scalable services. For example, PayPal is a java and spring-based application using Node.js.

Advantages of Node.js

The node.js applications are faster than the other framework-based applications and require fewer people to build the app.
It requires fewer lines of code.
The Node app has a 35% faster response time than the other apps.
The major advantage of using node.js is that node.js uses javascript. If you are a front-end developer, then you can easily transit from the front-end to the full stack developer.
Angular.js
Angular.js
is a JavaScript
framework that is used to develop web applications. This framework is developed by the Google. Now, the question arises that there are many javascript frameworks available in the market. Why do we prefer angular.js over the other frameworks for developing the web application?.

Advantages of Angular.js

It is a two-way data binding which means that it keeps the model and view in sync. If any changes are made in the model, then automatically view will also be updated accordingly.
The Angular.js is designed with testing in mind. The components of angular.js application can be tested with both the testing, such as unit testing and end to end testing.
With the help of Angular.js, it is easy to develop the application in an MVC architecture.
MongoDB
MongoDB
is the database used in web development. It is a NoSQL database, and a NoSQL database
can be defined as a non-relational and document-oriented database management system. As it is a document-oriented database management system, so it stores the data in the form of documents. The SQL databases use SQL query language to query the database, whereas the MongoDB is a NoSQL database that uses BSON language to query the database. JSON is a text-based format, but it is inefficient in terms of speed and space. In order to make MongoDB efficient in terms of space and speed, BSON was invented. BSON basically stores the JSON format in the binary form that optimizes the space, speed, and complexity.

Since the MongoDB is an unstructured schema and the data is not related to each other then the question arises 'why do we need to use the MongoDB?'. MongoDB is mainly useful for projects which are huge.

Express.js
Express.js
is a free and open-source software used as a back-end web application framework. It is commonly used in the popular development stacks like MEAN with a MongoDB database. The Express.js was developed by TJ Holowaychuk.